requirements.txt
astroid==2.11.3
atomicwrites==1.4.0
attrs==21.4.0
click==8.1.2
colorama==0.4.4
dill==0.3.4
Flask==2.1.1
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pylint==2.13.7
pyparsing==3.0.8
pytest==7.1.1
tomli==2.0.1
Werkzeug==2.1.1
wrapt==1.14.0
gunicorn==20.1.0
pymongo~=4.1.1
python-dotenv~=0.20.0
dnspython==2.2.1